refactor: align card layout  on dashboard
